Switch std::call_once to llvm::call_once
The std::call_once implementation in libstdc++ has problems on few systems:
NetBSD, OpenBSD and Linux PPC. LLVM ships with a homegrown implementation
llvm::call_once to help on these platforms.

This change is as of now required in LLDB as well to help the NetBSD port.

For consistency swich all std::call_once to llvm::call_once.